requirements.txt
pyserial==3.5
modbus-tk==1.1.2
tqdm==4.62.3
